requirements.txt
torch
transformers
flask
flask-restful
flask-cors
pillow
protobuf
bitsandbytes
optimum
accelerate